// src/gen-ts/src/models/YearlyReport.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Company Management API
 * API CRUD pour la gestion des entreprises, chantiers, entrepôts, équipements, matériaux, mouvements financiers, dépenses, achats et frais de transport. Basée sur le schéma de base de données fourni.
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime'
import type { JobWithFinancials } from './JobWithFinancials'
import {
  JobWithFinancialsFromJSON,
  JobWithFinancialsFromJSONTyped,
  JobWithFinancialsToJSON,
  JobWithFinancialsToJSONTyped,
} from './JobWithFinancials'
import type { YearlyReportSummary } from './YearlyReportSummary'
import {
  YearlyReportSummaryFromJSON,
  YearlyReportSummaryFromJSONTyped,
  YearlyReportSummaryToJSON,
  YearlyReportSummaryToJSONTyped,
} from './YearlyReportSummary'

/**
 *
 * @export
 * @interface YearlyReport
 */
export interface YearlyReport {
  /**
   *
   * @type {number}
   * @memberof YearlyReport
   */
  year?: number
  /**
   *
   * @type {Array<JobWithFinancials>}
   * @memberof YearlyReport
   */
  jobs_with_financials?: Array<JobWithFinancials>
  /**
   *
   * @type {YearlyReportSummary}
   * @memberof YearlyReport
   */
  summary?: YearlyReportSummary
}

/**
 * Check if a given object implements the YearlyReport interface.
 */
export function instanceOfYearlyReport(value: object): value is YearlyReport {
  return true
}

export function YearlyReportFromJSON(json: any): YearlyReport {
  return YearlyReportFromJSONTyped(json, false)
}

export function YearlyReportFromJSONTyped(json: any, ignoreDiscriminator: boolean): YearlyReport {
  if (json == null) {
    return json
  }
  return {
    year: json['year'] == null ? undefined : json['year'],
    jobs_with_financials:
      json['jobs_with_financials'] == null
        ? undefined
        : (json['jobs_with_financials'] as Array<any>).map(JobWithFinancialsFromJSON),
    summary: json['summary'] == null ? undefined : YearlyReportSummaryFromJSON(json['summary']),
  }
}

export function YearlyReportToJSON(json: any): YearlyReport {
  return YearlyReportToJSONTyped(json, false)
}

export function YearlyReportToJSONTyped(
  value?: YearlyReport | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value
  }

  return {
    year: value['year'],
    jobs_with_financials:
      value['jobs_with_financials'] == null
        ? undefined
        : (value['jobs_with_financials'] as Array<any>).map(JobWithFinancialsToJSON),
    summary: YearlyReportSummaryToJSON(value['summary']),
  }
}