// src/gen-ts/src/models/YearlyReportSummary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Company Management API
 * API CRUD pour la gestion des entreprises, chantiers, entrepôts, équipements, matériaux, mouvements financiers, dépenses, achats et frais de transport. Basée sur le schéma de base de données fourni.
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime'
/**
 *
 * @export
 * @interface YearlyReportSummary
 */
export interface YearlyReportSummary {
  /**
   *
   * @type {number}
   * @memberof YearlyReportSummary
   */
  total_income?: number
  /**
   *
   * @type {number}
   * @memberof YearlyReportSummary
   */
  total_expense?: number
  /**
   *
   * @type {number}
   * @memberof YearlyReportSummary
   */
  net_profit?: number
  /**
   *
   * @type {number}
   * @memberof YearlyReportSummary
   */
  job_count?: number
  /**
   *
   * @type {number}
   * @memberof YearlyReportSummary
   */
  in_progress_job_count?: number
}

/**
 * Check if a given object implements the YearlyReportSummary interface.
 */
export function instanceOfYearlyReportSummary(value: object): value is YearlyReportSummary {
  return true
}

export function YearlyReportSummaryFromJSON(json: any): YearlyReportSummary {
  return YearlyReportSummaryFromJSONTyped(json, false)
}

export function YearlyReportSummaryFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): YearlyReportSummary {
  if (json == null) {
    return json
  }
  return {
    total_income: json['total_income'] == null ? undefined : json['total_income'],
    total_expense: json['total_expense'] == null ? undefined : json['total_expense'],
    net_profit: json['net_profit'] == null ? undefined : json['net_profit'],
    job_count: json['job_count'] == null ? undefined : json['job_count'],
    in_progress_job_count:
      json['in_progress_job_count'] == null ? undefined : json['in_progress_job_count'],
  }
}

export function YearlyReportSummaryToJSON(json: any): YearlyReportSummary {
  return YearlyReportSummaryToJSONTyped(json, false)
}

export function YearlyReportSummaryToJSONTyped(
  value?: YearlyReportSummary | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value
  }

  return {
    total_income: value['total_income'],
    total_expense: value['total_expense'],
    net_profit: value['net_profit'],
    job_count: value['job_count'],
    in_progress_job_count: value['in_progress_job_count'],
  }
}