// src/gen-ts/src/models/JobWithFinancials.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Company Management API
 * API CRUD pour la gestion des entreprises, chantiers, entrepôts, équipements, matériaux, mouvements financiers, dépenses, achats et frais de transport. Basée sur le schéma de base de données fourni. 
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Job } from './Job';
import {
    JobFromJSON,
    JobFromJSONTyped,
    JobToJSON,
    JobToJSONTyped,
} from './Job';

/**
 * 
 * @export
 * @interface JobWithFinancials
 */
export interface JobWithFinancials {
    /**
     * 
     * @type {Job}
     * @memberof JobWithFinancials
     */
    job?: Job;
    /**
     * Total income amount for this job
     * @type {number}
     * @memberof JobWithFinancials
     */
    total_income?: number;
    /**
     * Total expense amount for this job
     * @type {number}
     * @memberof JobWithFinancials
     */
    total_expense?: number;
    /**
     * Net profit (total_income - total_expense)
     * @type {number}
     * @memberof JobWithFinancials
     */
    net_profit?: number;
}

/**
 * Check if a given object implements the JobWithFinancials interface.
 */
export function instanceOfJobWithFinancials(value: object): value is JobWithFinancials {
    return true;
}

export function JobWithFinancialsFromJSON(json: any): JobWithFinancials {
    return JobWithFinancialsFromJSONTyped(json, false);
}

export function JobWithFinancialsFromJSONTyped(json: any, ignoreDiscriminator: boolean): JobWithFinancials {
    if (json == null) {
        return json;
    }
    return {
        
        'job': json['job'] == null ? undefined : JobFromJSON(json['job']),
        'total_income': json['total_income'] == null ? undefined : json['total_income'],
        'total_expense': json['total_expense'] == null ? undefined : json['total_expense'],
        'net_profit': json['net_profit'] == null ? undefined : json['net_profit'],
    };
}

export function JobWithFinancialsToJSON(json: any): JobWithFinancials {
    return JobWithFinancialsToJSONTyped(json, false);
}

export function JobWithFinancialsToJSONTyped(value?: JobWithFinancials | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'job': JobToJSON(value['job']),
        'total_income': value['total_income'],
        'total_expense': value['total_expense'],
        'net_profit': value['net_profit'],
    };
}